+++ libs/request/src/CommonMediaResponse.ts
@@ -1,6 +1,5 @@
+import type { HttpResponse } from '@svta/cml-utils'
 import type { CommonMediaRequest } from './CommonMediaRequest.ts'
-import type { ResourceTiming } from './ResourceTiming.ts'
-import type { ResponseTypeMap } from './ResponseTypeMap.ts'
 
 /**
  * Common response API.
@@ -8,49 +7,4 @@ import type { ResponseTypeMap } from './ResponseTypeMap.ts'
  * @public
  */
 
-export type CommonMediaResponse<R extends CommonMediaRequest = CommonMediaRequest> = {
-	/**
-	 * The origin request.
-	 */
-	request: R;
-
-	/**
-	 * The final URL obtained after any redirects.
-	 */
-	url?: string;
-
-	/**
-	 * Indicates whether or not the request was redirected.
-	 */
-	redirected?: boolean;
-
-	/**
-	 * The HTTP status code of the response.
-	 */
-	status?: number;
-
-	/**
-	 * The status message corresponding to the HTTP status code.
-	 */
-	statusText?: string;
-
-	/**
-	 * The type of the response.
-	 */
-	type?: string;
-
-	/**
-	 * The response headers.
-	 */
-	headers?: Record<string, string>;
-
-	/**
-	 * The response data.
-	 */
-	data?: ResponseTypeMap<R['responseType']>;
-
-	/**
-	 * The network timing of the request/response.
-	 */
-	resourceTiming?: ResourceTiming;
-};
+export type CommonMediaResponse<R extends CommonMediaRequest = CommonMediaRequest> = HttpResponse<R>;