+++ libs/cta/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.0.4] - 2026-02-04
+
+### Changed
+
+- Update `@svta/cml-utils` to 1.3.0
+- Update `@svta/cml-structured-field-values` to 1.1.1
+
 ## [1.0.3] - 2026-02-03
 
 ### Changed
@@ -35,7 +42,8 @@ and this project adheres to
 - Convert to mono-repo ([#238](https://github.com/streaming-video-technology-alliance/common-media-library/issues/238))
 - Produce single bundled export for each package ([#260](https://github.com/streaming-video-technology-alliance/common-media-library/issues/260))
 
-[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.3...HEAD
+[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.4...HEAD
+[1.0.4]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.3...cta-v1.0.4
 [1.0.3]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.2...cta-v1.0.3
 [1.0.2]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.1...cta-v1.0.2
 [1.0.1]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cta-v1.0.0...cta-v1.0.1