+++ libs/drm/CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.1.3] - 2026-02-04
+
+### Changed
+
+- Update `@svta/cml-iso-bmff` to 1.0.0-beta.2
+- Update `@svta/cml-utils` to 1.3.0
+- Update `@svta/cml-xml` to 1.1.2
+
 ## [1.1.2] - 2026-02-03
 
 ### Changed
@@ -46,7 +54,8 @@ and this project adheres to
 - Convert to mono-repo ([#238](https://github.com/streaming-video-technology-alliance/common-media-library/issues/238))
 - Produce single bundled export for each package ([#260](https://github.com/streaming-video-technology-alliance/common-media-library/issues/260))
 
-[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/drm-v1.1.2...HEAD
+[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/drm-v1.1.3...HEAD
+[1.1.3]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/drm-v1.1.2...drm-v1.1.3
 [1.1.2]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/drm-v1.1.1...drm-v1.1.2
 [1.1.1]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/drm-v1.1.0...drm-v1.1.1
 [1.1.0]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/drm-v1.0.2...drm-v1.1.0